+++ cloud_client.go
@@ -4,10 +4,10 @@ import (
 	"bytes"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"io"
 	"log"
 	"net/http"
+	"os"
 	"strings"
 	"time"
 )
@@ -35,6 +35,10 @@ func NewClient(config Config) *CloudClient {
 	timeout := config.TimeoutSecond
 	if timeout <= 0 {
 		timeout = 10
+	}
+
+	if config.CustomLogger == nil {
+		client.Config.CustomLogger = NewCustomLogger(INFO, os.Stdout)
 	}
 
 	client.HttpClient = &http.Client{
@@ -113,9 +117,21 @@ func (cloudClient *CloudClient) Request(method string, requestPath string, param
 		Headers(request, config.ApiKey, timestamp, sign, config.Headers)
 	}
 
-	if config.IsPrint {
-		fmt.Println("---------------------------------------------")
-		PrintRequest(request, jsonBody)
+	if config.CustomLogger.logLevel == DEBUG {
+		if method == "GET" {
+			config.CustomLogger.Logf(DEBUG, "[%s] url=%s",
+				request.Method,
+				request.URL.String(),
+			)
+		} else {
+			config.CustomLogger.Logf(DEBUG, "[%s] url=%s\n\tHeader: %s\n\tBody: %s\n",
+				request.Method,
+				request.URL.String(),
+				request.Header,
+				jsonBody,
+			)
+		}
+
 	}
 
 	// send a request to remote server, and get a response
@@ -142,6 +158,12 @@ func (cloudClient *CloudClient) Request(method string, requestPath string, param
 	cloudResponse.Limit.Remaining = StringToInt(response.Header.Get("X-BM-RateLimit-Remaining"))
 	cloudResponse.Limit.Reset = StringToInt(response.Header.Get("X-BM-RateLimit-Reset"))
 	cloudResponse.Limit.Mode = response.Header.Get("X-BM-RateLimit-Mode")
+
+	if config.CustomLogger.logLevel == DEBUG {
+		config.CustomLogger.Logf(DEBUG, "Response=%s\n",
+			cloudResponse.Response,
+		)
+	}
 
 	return response, nil
 }